feat: added social authentication, core upgrade improvement and cache…

+++ database/migrations/2023_01_01_000004_create_post_meta_table.php
@@ -12,6 +12,10 @@ return new class () extends Migration {
      */
     public function up(): void
     {
+        if (Schema::hasTable('post_meta')) {
+            return;
+        }
+
         Schema::create('post_meta', function (Blueprint $table) {
             $table->id();
             $table->foreignId('post_id')->constrained()->onDelete('cascade');
