+++ database/migrations/2025_08_13_153546_create_user_meta_table.php
@@ -10,6 +10,10 @@ return new class () extends Migration {
      */
     public function up(): void
     {
+        if (Schema::hasTable('user_meta')) {
+            return;
+        }
+
         Schema::create('user_meta', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
